refactor: split common.rs into approval_manager and response_builders (closes #474)
Extract ApprovalManager/CancelToken/CancelManager into approval_manager.rs
and response builder helpers into response_builders.rs. Re-export via
common.rs to keep all call sites unchanged.

common.rs: 1134 → 810 lines (−324)
approval_manager.rs: 292 lines (new, includes tests)
response_builders.rs: 48 lines (new)
